+++ GuaranteedRate.Sextant/Config/IniConfig.cs
@@ -8,13 +8,18 @@ using System.Threading.Tasks;
 
 namespace GuaranteedRate.Sextant.Config
 {
-    /**
-     * IniConfig is a simple config class that expects data in the old .INI format:
-     * [key]=[value]\n
-     * 
-     * Lines starting with # will be treated as comments
-     * Empty lines will be ignored.
-     */
+
+
+
+
+    /// <summary>
+    ///      IniConfig is a simple config class that expects data in the old .INI format:
+     /// [key]=[value]\n
+     /// 
+     /// Lines starting with # will be treated as comments
+     /// Empty lines will be ignored.
+    ///
+    /// </summary>
     public class IniConfig : IEncompassConfig
     {
         private volatile IDictionary<string, string> _config;
@@ -84,8 +89,8 @@ namespace GuaranteedRate.Sextant.Config
                 if (_config.TryGetValue(key.ToLower(), out retVal))
                 {
                     return retVal;
-                } 
-                else 
+                }
+                else
                 {
                     return defaultVal;
                 }
@@ -117,5 +122,15 @@ namespace GuaranteedRate.Sextant.Config
             }
         }
 
+
+        /// <summary>
+        /// Do not use.  Always returns null in the iniConfig impllementation
+        /// </summary>
+        /// <param name="key"></param>
+        /// <returns>null.  Always.</returns>
+        public IEncompassConfig GetConfigGroup(string key)
+        {
+            return null;
+        }
     }
 }