+++ GuaranteedRate.SextantTests/Properties/AssemblyInfo.cs
@@ -6,11 +6,11 @@ using System.Runtime.InteropServices;
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
 
-[assembly: AssemblyTitle("GuaranteedRate.Examples.UserUtils")]
+[assembly: AssemblyTitle("GuaranteedRate.SextantTests")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Microsoft")]
-[assembly: AssemblyProduct("GuaranteedRate.Examples.UserUtils")]
+[assembly: AssemblyProduct("GuaranteedRate.SextantTests")]
 [assembly: AssemblyCopyright("Copyright © Microsoft 2016")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -23,7 +23,7 @@ using System.Runtime.InteropServices;
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 
-[assembly: Guid("47dd2ec4-99f6-40cd-bdb7-15d15b72f124")]
+[assembly: Guid("3d92ab6d-16cd-4bdc-b64c-4aebefdac378")]
 
 // Version information for an assembly consists of the following four values:
 //